// resources/ts/features/menu/types/SurveyingList.ts
type surveyingListState = {
    surveyingList: surveyingList[] | null;
    removeModal: removeModal;
    searchValue: string;
    sortSurveyingList: surveyingList[] | null;
    loading: boolean;
    error: string | null | undefined;
};

type surveyingList = {
    surveyingListId: number;
    surveyingName: string;
    weather: number;
    author: number; //作成者、site_member_id
    authorName: string;
    createdAt: string;
};

type removeModal = {
    isOpen: boolean;
    removeSurveyingListId: number;
};

export type { surveyingListState, surveyingList };